Move QueueDriverEnum to ert.config
