+++ src/ert/job_queue/driver.py
@@ -1,29 +1,9 @@
 from typing import List, Optional, Tuple
 
-from cwrap import BaseCClass, BaseCEnum
+from cwrap import BaseCClass
 
 from ert._c_wrappers import ResPrototype
-
-
-class QueueDriverEnum(BaseCEnum):  # type: ignore
-    TYPE_NAME = "queue_driver_enum"
-    NULL_DRIVER = None
-    LSF_DRIVER = None
-    LOCAL_DRIVER = None
-    TORQUE_DRIVER = None
-    SLURM_DRIVER = None
-
-
-QueueDriverEnum.addEnum("NULL_DRIVER", 0)
-QueueDriverEnum.addEnum("LSF_DRIVER", 1)
-QueueDriverEnum.addEnum("LOCAL_DRIVER", 2)
-QueueDriverEnum.addEnum("TORQUE_DRIVER", 4)
-QueueDriverEnum.addEnum("SLURM_DRIVER", 5)
-
-
-LSF_DRIVER = QueueDriverEnum.LSF_DRIVER
-LOCAL_DRIVER = QueueDriverEnum.LOCAL_DRIVER
-SLURM_DRIVER = QueueDriverEnum.SLURM_DRIVER
+from ert.config import QueueDriverEnum
 
 
 class Driver(BaseCClass):  # type: ignore
