+++ tests/unit_tests/ensemble_evaluator/conftest.py
@@ -8,11 +8,11 @@ from unittest.mock import Mock
 import pytest
 
 import ert.ensemble_evaluator
-from ert.config import QueueConfig
+from ert.config import QueueConfig, QueueDriverEnum
 from ert.ensemble_evaluator.config import EvaluatorServerConfig
 from ert.ensemble_evaluator.evaluator import EnsembleEvaluator
 from ert.ensemble_evaluator.snapshot import SnapshotBuilder
-from ert.job_queue import ExtJob, QueueDriverEnum
+from ert.job_queue import ExtJob
 from ert.load_status import LoadStatus
 
 from .ensemble_evaluator_utils import TestEnsemble
